Add CFAllocator & CFDate CoreFoundation APIs
Getting the abstractions for CFAllocator right is going to be very
important as it forms such a big part of many other types found within
the Apple frameworks.

To ensure things are off to a reasonable start - CFDate bindings have
been added along with tests to ensure that they behave as expected,
which is a simple way to ensure that the underlying allocator is
behaving as expected.

The core/difficult problem this change tries to solve is adding an
"as-simple-as-possible" mechanism constructing a CFAllocator from a zig
allocator. As making a CFAllocator requires allocations, it's probably
best to have our abstractions take CFAllocators directly - requiring
users of the zig foundation abstractions to create a CFAllocator
up-front & re-use it when needed.